Move scoring logic into internal/scorer so it can be reused.
