+++ internal/scorer/algorithm/distribution.go
