+++ internal/scorer/algorithm/input.go
