+++ internal/scorer/algorithm/registry.go
